system: Move nvmf_tgt conf read functions to separate file
  - introduce SPDK_NO_NVMF_DIRECT to disable support for NVME Direct
