+++ app/nvmf_tgt/Makefile
@@ -40,7 +40,7 @@ APP = nvmf_tgt
 
 CFLAGS += $(ENV_CFLAGS)
 
-C_SRCS := conf.c nvmf_main.c nvmf_tgt.c nvmf_rpc.c
+C_SRCS := conf.c conf_read.c nvmf_main.c nvmf_tgt.c nvmf_rpc.c
 
 SPDK_LIB_LIST = nvmf event log trace conf util bdev copy rpc jsonrpc json
 SPDK_LIB_LIST += app_rpc log_rpc bdev_rpc
